[vector_scan_generic prod] bandaid for the problem of NaNs
In a cumulative product sequence, once we hit a zero, we should continue to emit zero indefinitely.
However, in a parallel scan, we compute partial products to the right of the 0 that might diverge
to floating point infinity (which is not infinity in a strict mathematical sense). When we then
propagate the zeros rightward, the infinities produces NaN results. Under the assumption that input
data contains only finite numbers, we should then treat 0s on the left as `absorbing`: regardless of
the value on the right, we emit 0.